+++ CoreVisualizer/ECVIGenericCPUCore.h
@@ -32,6 +32,6 @@
 @protocol ECVIGenericCPUCoreDelegate <NSObject>
 
 @optional
-- (void)CPUcore:(ECVIGenericCPUCore *)core didUpdateRegister:(uint32_t)rnum toValue:(uint64_t)value;
+- (void)CPUcore:(ECVIGenericCPUCore *)core didUpdateRegister:(uint32_t)rnum toValue:(uint128_t)value;
 
 @end
